+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 * Add `ruby-3.1` in CI
+* Add parsing `workbookProtection` in xlsx
 
 ### Changes